🐛 Fix Next.js "use server" directive conflict with sync function
Move categorizeService pure function and types to separate integration-utils.ts
file. Next.js "use server" files can only export async server actions, so the
pure synchronous categorizeService function was causing build failures.
